:sparkles: add serialize support
