+++ docs/changelog.md
@@ -7,11 +7,14 @@
   - 新增 `weekday` 和 `isoweekday` 方法
   - 补充公历2101年日期的干支、节气信息
   - 修正无节气的日期格式化的bug
+  - 新增 sqlte3 自定义字段支持
 - `calendars.festivals` 模块
   - 新增 `encode` / `decode` 方法，支持序列化
   - 新增 `FestivalFactory` ，支持节日分组分类
   - 支持省略年份的字符表达式
   - 修正 `SolarSchema` 二月倒数序号形式解析的bug
+- `calendars.birthday` 模块
+  - 新增虚岁/周岁的计算函数
 - 新增 `borax.structures.proxy` 模块
 - `trim_iterable` 新增前缀、后缀字符串
